Scale how the data is downloaded in get_data.py
If there are more than 1 pod, the init-container of one pod should download
the data while the init-containers of the other pods wait. Once the data is downloaded
the processing container of all the pods can access the data.

Update pman and job template
Change the memory value to 150Mi as that is the minimum allowed on MOC.
Change the volume type for the shared directory to be a persistent volume claim
with the name of the claim being `job_id`-storage-claim.

Add function to create and delete PVC for job
Creates a Persistent Volume Claim with the name "job-id"-storage-claim
when an OpenShift job is created.
Deletes the created PVC when the job is shut down.

Remove hardcoding for incoming and outgoing directories
Pick up the paths of the incoming and outgoing directories from
the cmd.

Signed-off-by: umohnani8 <umohnani@redhat.com>